Softaware/Sist_Avaliacao_IA/services/auth/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]>=2.0
asyncpg
passlib[bcrypt]
python-jose
python-multipart
python-dotenv
aiosmtplib
email-validator